<sge>     - minor update         - added macro SGE_STATIC_ASSERT_NO_MEMBER_CLASS         - naming convension: rename struct to class         - StringT added explicit bool operator     - gmake_win32 tool exe embed into repo         - use $info instead of @echo in GNUmakefile     -

+++ src/render/src/sge_render/mesh/RenderTerrain.cpp
@@ -223,7 +223,7 @@ void RenderTerrain::Patch::render(RenderRequest& req) {
 	_material->setParam("patchIndex", _index);
 
 	auto passes = _material->passes();
-	for (size_t i = 0; i < passes.size(); i++) {
+	for (int i = 0; i < passes.size(); i++) {
 		auto* cmd				= req.addDrawCall();
 #if _DEBUG
 		cmd->debugLoc			= SGE_LOC;
